Studio - Move leaf modules into src/apps/studio
Relocates studioTypes, studioTags, studioFieldMeta, studioParsley and FieldIconChip out of content-editor's ItemEdit view and into src/apps/studio. Import paths only; no behaviour change.
