+++ src/apps/studio/components/FieldIconChip.tsx
